requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.1
Django==4.0.6
django-environ==0.9.0
django-taggit==3.0.0
gunicorn==20.1.0
idna==3.3
Pillow==9.2.0
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.12
whitenoise==6.2.0
